Refactor input symbols component and rename it to qualifications
